changed the id field in ssmem_ts_t from uint8_t to size_t (in any case it would be aligned to 8 bytes)

+++ include/ssmem.h
@@ -65,7 +65,7 @@ typedef struct ALIGNED(CACHE_LINE_SIZE) ssmem_ts
     struct
     {
       size_t version;
-      uint8_t id;
+      size_t id;
       struct ssmem_ts* next;
     };
   };
